move download_modal into subdirectory
