+++ frontend/javascripts/viewer/view/action_bar/view_dataset_actions_view.tsx
@@ -20,7 +20,7 @@ import ShareViewDatasetModalView from "viewer/view/action_bar/share_view_dataset
 import ButtonComponent from "viewer/view/components/button_component";
 import { downloadScreenshot } from "viewer/view/rendering_utils";
 import CreateAnimationModal from "./create_animation_modal";
-import DownloadModalView from "./download_modal_view";
+import DownloadModalView from "./download_modal/download_modal_view";
 
 type Props = {
   layoutMenu: SubMenuType;
